Conditionally set expectations for fname restrictions based on db version

+++ tests/test_mongodb_fname_restrictions.py
@@ -6,7 +6,7 @@ from bson import ObjectId
 
 from mongodb_common import drop_all_collections, ensure_environment_variables_set, \
     get_test_connection
-from tap_tester import connections, menagerie, runner
+from tap_tester import connections, menagerie, runner, LOGGER
 
 
 RECORD_COUNT = {}
@@ -40,6 +40,10 @@ class MongoDBFieldNameRestrictions(unittest.TestCase):
         ensure_environment_variables_set()
 
         with get_test_connection() as client:
+
+            self.db_version = client.server_info()['version']
+            LOGGER.info("MongoDB version {}".format(self.db_version))
+
             ############# Drop all dbs/collections #############
             drop_all_collections(client)
 
@@ -50,36 +54,71 @@ class MongoDBFieldNameRestrictions(unittest.TestCase):
             # simple_coll_2 has 100 documents
             client["simple_db"]["simple_coll_2"].insert_many(generate_simple_coll_docs(100))
 
-            # simple_coll_3 has 10 documents
-            client["simple_db"]["simple_coll_3"].insert_many(generate_leading_coll_docs(10))
+            if self.db_version == '4.4.6':
+                with self.assertRaises(Exception) as e:
+                    # simple_coll_3 has 10 documents
+                    client["simple_db"]["simple_coll_3"].insert_many(generate_leading_coll_docs(10))
+
+                e_error = "Document can't have $ prefixed field names"
+                self.assertIn('pymongo.errors.BulkWriteError', str(e.exception.__class__))
+                self.assertIn(e_error, str(e.exception.details))
+
+            else:
+                # simple_coll_3 has 10 documents
+                client["simple_db"]["simple_coll_3"].insert_many(generate_leading_coll_docs(10))
+
 
     def expected_check_streams(self):
-        return {
-            'simple_db-simple_coll_1',
-            'simple_db-simple_coll_2',
-            'simple_db-simple_coll_3',
-        }
+        if self.db_version == '4.4.6':
+            return {
+                'simple_db-simple_coll_1',
+                'simple_db-simple_coll_2',
+            }
+        else:
+            return {
+                'simple_db-simple_coll_1',
+                'simple_db-simple_coll_2',
+                'simple_db-simple_coll_3',
+            }
 
     def expected_pks(self):
-        return {
-            'simple_coll_1': {'_id'},
-            'simple_coll_2': {'_id'},
-            'simple_coll_3': {'_id'},
-        }
+        if self.db_version == '4.4.6':
+            return {
+                'simple_coll_1': {'_id'},
+                'simple_coll_2': {'_id'},
+            }
+        else:
+            return {
+                'simple_coll_1': {'_id'},
+                'simple_coll_2': {'_id'},
+                'simple_coll_3': {'_id'},
+            }
 
     def expected_row_counts(self):
-        return {
-            'simple_coll_1': 50,
-            'simple_coll_2': 100,
-            'simple_coll_3': 10,
-        }
+        if self.db_version == '4.4.6':
+            return {
+                'simple_coll_1': 50,
+                'simple_coll_2': 100,
+            }
+        else:
+            return {
+                'simple_coll_1': 50,
+                'simple_coll_2': 100,
+                'simple_coll_3': 10,
+            }
 
     def expected_sync_streams(self):
-        return {
-            'simple_coll_1',
-            'simple_coll_2',
-            'simple_coll_3',
-        }
+        if self.db_version == '4.4.6':
+            return {
+                'simple_coll_1',
+                'simple_coll_2',
+            }
+        else:
+            return {
+                'simple_coll_1',
+                'simple_coll_2',
+                'simple_coll_3',
+            }
 
     def name(self):
         return "tap_tester_mongodb_fname_restrict"
@@ -200,13 +239,14 @@ class MongoDBFieldNameRestrictions(unittest.TestCase):
             changed_ids.add(object_id)
             client["simple_db"]["simple_coll_2"].delete_one({'_id': object_id})
 
-            object_id = client['simple_db']['simple_coll_3'].find()[0]['_id']
-            changed_ids.add(object_id)
-            client["simple_db"]["simple_coll_3"].delete_one({'_id': object_id})
+            if self.db_version != '4.4.6':
+                object_id = client['simple_db']['simple_coll_3'].find()[0]['_id']
+                changed_ids.add(object_id)
+                client["simple_db"]["simple_coll_3"].delete_one({'_id': object_id})
 
-            object_id = client['simple_db']['simple_coll_3'].find()[1]['_id']
-            changed_ids.add(object_id)
-            client["simple_db"]["simple_coll_3"].delete_one({'_id': object_id})
+                object_id = client['simple_db']['simple_coll_3'].find()[1]['_id']
+                changed_ids.add(object_id)
+                client["simple_db"]["simple_coll_3"].delete_one({'_id': object_id})
 
             # Update two documents for each collection
 
@@ -245,30 +285,31 @@ class MongoDBFieldNameRestrictions(unittest.TestCase):
             client["simple_db"]["simple_coll_2"].update_one(
                 {'_id': object_id},{'$set': {'string$field': "Updated_2"}})
 
-            num_records = len(list(client['simple_db']['simple_coll_3'].find()))
-            last_index = num_records - 1
-            sec_last_index = num_records - 2
+            if self.db_version != '4.4.6':
+                num_records = len(list(client['simple_db']['simple_coll_3'].find()))
+                last_index = num_records - 1
+                sec_last_index = num_records - 2
 
-            object_id = client['simple_db']['simple_coll_3'].find()[sec_last_index]['_id']
-            # changed_ids.add(object_id)  # this results in empty field and does not replicate
-            with self.assertRaises(Exception) as e1:
-                client["simple_db"]["simple_coll_3"].update_one(
-                    {'_id': object_id},{'$set': {'.int_field': -1}})
+                object_id = client['simple_db']['simple_coll_3'].find()[sec_last_index]['_id']
+                # changed_ids.add(object_id)  # this results in empty field and does not replicate
+                with self.assertRaises(Exception) as e1:
+                    client["simple_db"]["simple_coll_3"].update_one(
+                        {'_id': object_id},{'$set': {'.int_field': -1}})
 
-            e1_error = ".int_field' contains an empty field name, which is not allowed"
-            self.assertIn('pymongo.errors.WriteError', str(e1.exception.__class__))
-            self.assertIn(e1_error, str(e1.exception.details))
+                    e1_error = ".int_field' contains an empty field name, which is not allowed"
+                    self.assertIn('pymongo.errors.WriteError', str(e1.exception.__class__))
+                    self.assertIn(e1_error, str(e1.exception.details))
 
-            object_id = client['simple_db']['simple_coll_3'].find()[last_index]['_id']
-            changed_ids.add(object_id)
-            with self.assertRaises(Exception) as e2:
-                client["simple_db"]["simple_coll_3"].update_one(
-                    {'_id': object_id},{'$set': {'$string_field': "Updated_3"}})
+                    object_id = client['simple_db']['simple_coll_3'].find()[last_index]['_id']
+                    changed_ids.add(object_id)
+                    with self.assertRaises(Exception) as e2:
+                        client["simple_db"]["simple_coll_3"].update_one(
+                            {'_id': object_id},{'$set': {'$string_field': "Updated_3"}})
 
-            e2_error = ( "The dollar ($) prefixed field '$string_field' in '$string_field' is not "
-                         "allowed in the context of an update's replacement document." )
-            self.assertIn('pymongo.errors.WriteError', str(e2.exception.__class__))
-            self.assertIn(e2_error, str(e2.exception.details))
+                        e2_error = ( "The dollar ($) prefixed field '$string_field' in '$string_field' is not "
+                                     "allowed in the context of an update's replacement document." )
+                        self.assertIn('pymongo.errors.WriteError', str(e2.exception.__class__))
+                        self.assertIn(e2_error, str(e2.exception.details))
 
             # Insert two documents for each collection
             last_index = len(list(client['simple_db']['simple_coll_1'].find()))
@@ -295,17 +336,18 @@ class MongoDBFieldNameRestrictions(unittest.TestCase):
             object_id = client["simple_db"]["simple_coll_2"].find()[last_index]['_id']
             changed_ids.add(object_id)
 
-            last_index = len(list(client['simple_db']['simple_coll_3'].find()))
-            client["simple_db"]["simple_coll_3"].insert_one(
-                {".int_field": 300, "$string_field": random_string_generator()})
-            object_id = client["simple_db"]["simple_coll_3"].find()[last_index]['_id']
-            changed_ids.add(object_id)
+            if self.db_version != '4.4.6':
+                last_index = len(list(client['simple_db']['simple_coll_3'].find()))
+                client["simple_db"]["simple_coll_3"].insert_one(
+                    {".int_field": 300, "$string_field": random_string_generator()})
+                object_id = client["simple_db"]["simple_coll_3"].find()[last_index]['_id']
+                changed_ids.add(object_id)
 
-            last_index += 1
-            client["simple_db"]["simple_coll_3"].insert_one(
-                {".int_field": 301, "$string_field": random_string_generator()})
-            object_id = client["simple_db"]["simple_coll_3"].find()[last_index]['_id']
-            changed_ids.add(object_id)
+                last_index += 1
+                client["simple_db"]["simple_coll_3"].insert_one(
+                    {".int_field": 301, "$string_field": random_string_generator()})
+                object_id = client["simple_db"]["simple_coll_3"].find()[last_index]['_id']
+                changed_ids.add(object_id)
 
         #  -------------------------------------------
         #  ----------- Subsequent Oplog Sync ---------
@@ -340,14 +382,17 @@ class MongoDBFieldNameRestrictions(unittest.TestCase):
                                  if x['data'].get('_sdc_deleted_at')]))
         self.assertEqual(2, len([x['data'] for x in records_by_stream['simple_coll_2']
                                  if x['data'].get('_sdc_deleted_at')]))
-        self.assertEqual(2, len([x['data'] for x in records_by_stream['simple_coll_3']
-                                 if x['data'].get('_sdc_deleted_at')]))
+        if self.db_version != '4.4.6':
+            self.assertEqual(2, len([x['data'] for x in records_by_stream['simple_coll_3']
+                                     if x['data'].get('_sdc_deleted_at')]))
 
         # Verify that the _id of the records sent are the same set as the
         # _ids of the documents changed
         actual = set([ObjectId(x['data']['_id']) for x in records_by_stream['simple_coll_1']]).union(
-            set([ObjectId(x['data']['_id']) for x in records_by_stream['simple_coll_2']])).union(
-                set([ObjectId(x['data']['_id']) for x in records_by_stream['simple_coll_3']]))
+            set([ObjectId(x['data']['_id']) for x in records_by_stream['simple_coll_2']]))
+        if self.db_version != '4.4.6':
+            actual.union(set([ObjectId(x['data']['_id']) for x
+                              in records_by_stream['simple_coll_3']]))
         self.assertEqual(changed_ids, actual)
 
         with get_test_connection() as client:
